# stroom-app/src/main/resources/ui/noauth/swagger/stroom.yaml
openapi: 3.0.1
info:
  contact:
    name: stroom
    url: https://github.com/gchq/stroom
  description: Various APIs for interacting with Stroom and its data
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Stroom API
  version: v1/v2
servers:
- url: /api
security:
- ApiKeyAuth: []
paths:
...
  /config/v1:
    post:
      operationId: createConfigProperty
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigProperty'
        description: configProperty
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigProperty'
          description: default response
      summary: Create a configuration property
      tags:
      - Global Config
  /config/v1/clusterProperties/{propertyName}:
    put:
      operationId: updateConfigProperty
      parameters:
      - in: path
        name: propertyName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigProperty'
        description: configProperty
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigProperty'
          description: default response
      summary: Update a configuration property
      tags:
      - Global Config
  /config/v1/clusterProperties/{propertyName}/yamlOverrideValue/{nodeName}:
    get:
      operationId: getConfigYamlValueByNodeAndName
      parameters:
      - in: path
        name: propertyName
        required: true
        schema:
          type: string
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverrideValueString'
          description: default response
      summary: Get the property value from the YAML configuration in the specified
        node.
      tags:
      - Global Config
...
  /config/v1/nodeProperties/{nodeName}:
    post:
      operationId: listConfigPropertiesByNode
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalConfigCriteria'
        description: criteria
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListConfigResponse'
          description: default response
      summary: List all properties matching the criteria on the requested node.
      tags:
      - Global Config
  /config/v1/properties:
    post:
      operationId: listConfigProperties
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalConfigCriteria'
        description: criteria
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListConfigResponse'
          description: default response
      summary: List all properties matching the criteria on the current node.
      tags:
      - Global Config
  /config/v1/properties/{propertyName}:
    get:
      operationId: getConfigPropertyByName
      parameters:
      - in: path
        name: propertyName
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigProperty'
          description: default response
      summary: Fetch a property by its name, e.g. 'stroom.path.home'
      tags:
      - Global Config
...
  /index/volume/v2:
    post:
      operationId: createIndexVolume
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexVolume'
        description: request
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexVolume'
          description: default response
      summary: Creates an index volume
      tags:
      - Index Volumes
  /index/volume/v2/find:
    post:
      operationId: findIndexVolumes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpressionCriteria'
        description: request
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageIndexVolume'
          description: default response
      summary: Finds index volumes matching request
      tags:
      - Index Volumes
  /index/volume/v2/rescan:
    delete:
      operationId: rescanIndexVolumes
      parameters:
      - in: query
        name: memberUuid
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: boolean
          description: default response
      summary: Rescans index volumes
      tags:
      - Index Volumes
  /index/volume/v2/{id}:
    delete:
      operationId: deleteIndexVolume
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        default:
          content:
            application/json:
              schema:
                type: boolean
          description: default response
      summary: Deletes an index volume
      tags:
      - Index Volumes
    get:
      operationId: fetchIndexVolume
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexVolume'
          description: default response
      summary: Fetch an index volume
      tags:
      - Index Volumes
    put:
      operationId: updateIndexVolume
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexVolume'
        description: indexVolume
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexVolume'
          description: default response
      summary: Updates an index volume
      tags:
      - Index Volumes
...
  /jobNode/v1:
    get:
      operationId: listJobsNodes
      parameters:
      - in: query
        name: jobName
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageJobNode'
          description: default response
      summary: Lists job nodes
      tags:
      - Jobs (Node)
  /jobNode/v1/info:
    get:
      operationId: fetchJobNodeInfo
      parameters:
      - in: query
        name: jobName
        schema:
          type: string
      - in: query
        name: nodeName
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobNodeInfo'
          description: default response
      summary: Gets current info for a job node
      tags:
      - Jobs (Node)
  /jobNode/v1/{id}/enabled:
    put:
      operationId: setJobNodeEnabled
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: boolean
        description: enabled
        required: true
      responses:
        default:
          content:
            application/json: {}
          description: default response
      summary: Sets the enabled status of the job node
      tags:
      - Jobs (Node)
  /jobNode/v1/{id}/schedule:
    put:
      operationId: setJobNodeSchedule
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: string
        description: schedule
        required: true
      responses:
        default:
          content:
            application/json: {}
          description: default response
      summary: Sets the schedule job node
      tags:
      - Jobs (Node)
  /jobNode/v1/{id}/taskLimit:
    put:
      operationId: setJobNodeTaskLimit
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int32
        description: taskLimit
        required: true
      responses:
        default:
          content:
            application/json: {}
          description: default response
      summary: Sets the task limit for the job node
      tags:
      - Jobs (Node)
...
components:
  schemas:
...
    ConfigProperty:
      type: object
      properties:
        createTimeMs:
          type: integer
          format: int64
        createUser:
          type: string
        dataTypeName:
          type: string
        databaseOverrideValue:
          $ref: '#/components/schemas/OverrideValueString'
        defaultValue:
          type: string
        description:
          type: string
        editable:
          type: boolean
        id:
          type: integer
          format: int32
        name:
          $ref: '#/components/schemas/PropertyPath'
        password:
          type: boolean
        requireRestart:
          type: boolean
        requireUiRestart:
          type: boolean
        updateTimeMs:
          type: integer
          format: int64
        updateUser:
          type: string
        version:
          type: integer
          format: int32
        yamlOverrideValue:
          $ref: '#/components/schemas/OverrideValueString'
...
    CriteriaFieldSort:
      type: object
      properties:
        desc:
          type: boolean
        id:
          type: string
        ignoreCase:
          type: boolean
...
    ExpressionCriteria:
      type: object
      properties:
        expression:
          $ref: '#/components/schemas/ExpressionOperator'
        pageRequest:
          $ref: '#/components/schemas/PageRequest'
        sort:
          type: string
          writeOnly: true
        sortList:
          type: array
          items:
            $ref: '#/components/schemas/CriteriaFieldSort'
    ExpressionItem:
      type: object
      description: Base type for an item in an expression tree
      discriminator:
        propertyName: type
      properties:
        enabled:
          type: boolean
          description: Whether this item in the expression tree is enabled or not
          example: true
        type:
          type: string
      required:
      - type
    ExpressionOperator:
      type: object
      description: A logical addOperator term in a query expression tree
      properties:
        children:
          type: array
          items:
            $ref: '#/components/schemas/ExpressionItem'
          xml:
            wrapped: true
        enabled:
          type: boolean
          description: Whether this item in the expression tree is enabled or not
          example: true
        op:
          type: string
          description: The logical addOperator type
          enum:
          - AND
          - OR
          - NOT
      required:
      - op
...
    GlobalConfigCriteria:
      type: object
      properties:
        pageRequest:
          $ref: '#/components/schemas/PageRequest'
        quickFilterInput:
          type: string
        sort:
          type: string
          writeOnly: true
        sortList:
          type: array
          items:
            $ref: '#/components/schemas/CriteriaFieldSort'
...
    IndexVolume:
      type: object
      properties:
        bytesFree:
          type: integer
          format: int64
        bytesLimit:
          type: integer
          format: int64
        bytesTotal:
          type: integer
          format: int64
        bytesUsed:
          type: integer
          format: int64
        createTimeMs:
          type: integer
          format: int64
        createUser:
          type: string
        id:
          type: integer
          format: int32
        indexVolumeGroupId:
          type: integer
          format: int32
        nodeName:
          type: string
        path:
          type: string
        state:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - CLOSED
        statusMs:
          type: integer
          format: int64
        updateTimeMs:
          type: integer
          format: int64
        updateUser:
          type: string
        version:
          type: integer
          format: int32
...
    Job:
      type: object
      properties:
        advanced:
          type: boolean
        createTimeMs:
          type: integer
          format: int64
        createUser:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        id:
          type: integer
          format: int32
        name:
          type: string
        updateTimeMs:
          type: integer
          format: int64
        updateUser:
          type: string
        version:
          type: integer
          format: int32
    JobNode:
      type: object
      properties:
        createTimeMs:
          type: integer
          format: int64
        createUser:
          type: string
        enabled:
          type: boolean
        id:
          type: integer
          format: int32
        job:
          $ref: '#/components/schemas/Job'
        jobType:
          type: string
          enum:
          - UNKNOWN
          - CRON
          - FREQUENCY
          - DISTRIBUTED
        nodeName:
          type: string
        schedule:
          type: string
        taskLimit:
          type: integer
          format: int32
        updateTimeMs:
          type: integer
          format: int64
        updateUser:
          type: string
        version:
          type: integer
          format: int32
    JobNodeInfo:
      type: object
      properties:
        currentTaskCount:
          type: integer
          format: int32
        lastExecutedTime:
          type: integer
          format: int64
        scheduleReferenceTime:
          type: integer
          format: int64
...
    ListConfigResponse:
      type: object
      description: List of config properties
      properties:
        nodeName:
          type: string
        pageResponse:
          $ref: '#/components/schemas/PageResponse'
        qualifiedFilterInput:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/ConfigProperty'
...
    OverrideValueString:
      type: object
      properties:
        hasOverride:
          type: boolean
        value:
          type: string
    PageRequest:
      type: object
      properties:
        length:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    PageResponse:
      type: object
      description: Details of the page of results being returned.
      properties:
        exact:
          type: boolean
        length:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
...
    PropertyPath:
      type: object
      properties:
        parts:
          type: array
          items:
            type: string
...
    ResultPageIndexVolume:
      type: object
      description: A page of results.
      properties:
        pageResponse:
          $ref: '#/components/schemas/PageResponse'
        values:
          type: array
          items:
            $ref: '#/components/schemas/IndexVolume'
...
    ResultPageJobNode:
      type: object
      description: A page of results.
      properties:
        pageResponse:
          $ref: '#/components/schemas/PageResponse'
        values:
          type: array
          items:
            $ref: '#/components/schemas/JobNode'
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey